Fix concurrency bug
If there's a @state in a (swap! state ... ) function, that's a smell.

It's the equivalent of the x = x + 1 threading bug where there are 3 steps:

1 read x
2 increment it
3 write x

If another thread reads x when your thread is in (2), the other thread will
see a wrong value.

Similarly, if you deref an atom with @state in a swap! function, you'll also
suffer from the 3 step pattern of tears, pain and death.